+++ src/main/java/com/example/dolbomi/admin/controller/dto/RegisterRequestDto.java
@@ -9,15 +9,14 @@ import javax.validation.constraints.NotNull;
 public class RegisterRequestDto {
 
     @NotNull(message = "노인 등록번호를 입력해주세요")
-    private Long registerNo;
+    private Integer registerNo;
 
     @NotNull(message = "관리자 인덱스를 입력해주세요")
-    private Long adminIndex;
+    private Integer adminIndex;
 
     @Builder
-    public RegisterRequestDto(Long registerNo, Long adminIndex) {
+    public RegisterRequestDto(Integer registerNo, Integer adminIndex) {
         this.registerNo = registerNo;
         this.adminIndex = adminIndex;
     }
-
 }
